move proxy server codebase into pkg/proxy
